#3 setup node server and sequelize database connection
