+++ .env.example
@@ -1,0 +1,7 @@
+PORT = 8000
+DATABASE_NAME = atabb
+DATABASE_USERNAME = geo
+DATABASE_PASSWORD = e808e9-
+HOST = localhost
+DATABASE_PORT = 5432
+
