+++ DATABASE.md
@@ -1,0 +1,7 @@
+### To test the database connection:
+
+\*\* Set up a PostgreSQL database either locally or online.
+
+\*\* Get your database credentials such as database name, database username, database password and database host.
+
+\*\* Create a .env file and input those credentials there using the fields DATABASE_NAME, DATABASE_USERNAME, DATABASE_PASSWORD, DATABASE_HOST as stated in the .env.example file.
